resort script: skip symlinks processing

+++ resort_stuff.sh
@@ -3,6 +3,7 @@
 for a in $(find -type d -name "package.*" ! -empty) "sets"; do
     cd "${a}"
     for b in *; do
+        [ -L "$b" ] && continue
         sort -u "${b}" > asd
         mv asd "${b}"
     done
